# Conflicts:
#	TP4/src/main/java/com/integrador4/Integrador4Application.java
#	TP4/src/main/java/com/integrador4/controller/ClientController.java
#	TP4/src/main/java/com/integrador4/controller/ProductController.java
#	TP4/src/main/java/com/integrador4/controller/SaleController.java
#	TP4/src/main/java/com/integrador4/dto/SaleProductDto.java
#	TP4/src/main/java/com/integrador4/entity/Client.java
#	TP4/src/main/java/com/integrador4/entity/Product.java
#	TP4/src/main/java/com/integrador4/entity/Sale.java
#	TP4/src/main/java/com/integrador4/entity/SaleProduct.java
#	TP4/src/main/java/com/integrador4/entity/VentaProductoPk.java
#	TP4/src/main/java/com/integrador4/repository/ClientRepo.java
#	TP4/src/main/java/com/integrador4/repository/ClientRepository.java
#	TP4/src/main/java/com/integrador4/repository/ClienteRepo.java
#	TP4/src/main/java/com/integrador4/repository/ProductRepository.java
#	TP4/src/main/java/com/integrador4/repository/SaleProductRepo.java
#	TP4/src/main/java/com/integrador4/repository/VentaProductoRepo.java
#	TP4/src/main/java/com/integrador4/repository/VentaProductoRepository.java
#	TP4/src/main/java/com/integrador4/service/ClientService.java
#	TP4/src/main/java/com/integrador4/service/ProductService.java
#	TP4/src/main/java/com/integrador4/service/SaleService.java

+++ TP4/src/main/java/com/integrador4/entity/Product.java
@@ -5,7 +5,6 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import javax.persistence.*;
-import java.math.BigDecimal;
 
 @Entity
 @Table( name = "product")
@@ -15,17 +14,21 @@ public class Product {
 
     @Id
     @GeneratedValue( strategy = GenerationType.AUTO )
-    private BigDecimal idProduct;
+    private Integer idProduct;
+
     @Column
     private String name;
+
     @Column
     private Double price;
-    @Column @JsonIgnore
+
+    @Column
+    @JsonIgnore
     private Integer stock;
 
     public Product() {}
 
-    public Product(BigDecimal id, String name, double price, int stock ){
+    public Product(Integer id, String name, double price, int stock ){
         this.idProduct = id;
         this.name = name;
         this.price = price;
@@ -38,7 +41,7 @@ public class Product {
         this.stock = stock;
     }
 
-    public BigDecimal getIdProduct() {
+    public Integer getIdProduct() {
         return idProduct;
     }
